linux-gen: modular-framework: move queue subsystem header to a public folder
Move queue subsystem header to a public folder.
